Tweaked informative region selection
The zeros from the normal cloud are now ignored. This should not remove the 'actual' zeros, since those should be 1e-10 or something similar most of the time, if they are taken from actual measurements or calculations.
Also, the code was pretty horrible at that section, so I rewrote it to look less like the work of a 1-year highschool student.