Added a non-functional PVM
The current implementation is based on the assumption that there cannot be two features at the exact same spacial location, but this does not hold true: it is possible to have a two features at the same coordinates, but with different scales or orientations.
The other assumption is that the values inside the matches list are unique, but this does not hold true either.